requirements.txt
colorama==0.4.6
tabulate==0.9.0
openai==1.3.0
python-dotenv==1.0.0
